optimization and cleanups
I removed the "user" global variable as there is no need to parse out the user in the logs for Sguil.  There isn't a Sguil DB field for user.  Therefore if OSSEC decodes it just add it to the body of the alert.  This makes it easier to identify the user account associated with the log when viewing the alert detail in Sguil as it will be on a separate line at the top of the alert detail.

I also made the processdata procedure a little easier for me to read.

In the processdata procedure I also replaced the ".*" for SRC IP and DST IP matching with "\S+".

Having an issue that some of my SRC and DST ports for alerts are not getting imported into Sguil even though they are being decoded in OSSEC and the ossec-agent.tcl script is parsing them out as they are not showing up in the Sguil alert detail.  Not sure what is causing this yet as some come into Sguil without issue.